+++ KFXUtilities/Classes/KFXCommonConstants.h
@@ -1,0 +1,26 @@
+
+
+#import <Foundation/Foundation.h>
+
+@interface KFXCommonConstants : NSObject
+
+
+//--------------------------------------------------------
+#pragma mark - GeoLocation Constants
+//--------------------------------------------------------
+extern double const kMetresPerDegreeLatitude;
+extern double const kMetresPerDegreeLongitude;
+extern double const kDegreesPerMetreLatitude;
+extern double const kDegreesPerMetreLongitude;
+
+extern double const kFeetPerMile;
+extern double const kMetresPerMile;
+extern double const kFeetPerKilometre;
+extern double const kKilometresPerMile;
+extern double const kMilesPerKilometre;
+
+
+
+
+
+@end