+++ KFXUtilities/Classes/KFXMaths.m
@@ -4,22 +4,6 @@
 #import "KFXMaths.h"
 
 
-//--------------------------------------------------------
-#pragma mark - GeoLocation Constants
-//--------------------------------------------------------
-double const kMetresPerDegreeLatitude = 110574.27608210722;
-double const kMetresPerDegreeLongitude = 111319.2895857656;
-double const kDegreesPerMetreLatitude = 0.0000090437;
-double const kDegreesPerMetreLongitude = 0.0000089832;
-
-
-double const kFeetPerMile = 5280.0;
-double const kMetresPerMile = 1609.34;
-double const kFeetPerKilometre = 3280.838879986877;
-double const kKilometresPerMile = 1.60934;
-double const kMilesPerKilometre = 0.621371;
-
-
 
 @implementation KFXMaths